+++ Classes/VGTaggerAppDelegate.m
@@ -30,4 +30,5 @@
 	[DDLog addLogger:[DDTTYLogger sharedInstance]];
 }
 
+
 @end